// src/auth/dto/payload-user.dto.ts
export class PayloadUserDto {
  name: string;
  title: string;
  uid: string;
  company: string;
  companyid: string;
  companyuserid: string;
  email: string;
  istrial: string;
  dateend: string;
  role: string;
  candownload: string;
  nbf: string;
  exp: string;
  iat: string;
  iss: string;
  aud: string;
}