// src/auth/dto/payload-company.dto.ts
export class PayloadCompanyDto {
  name: string;
  folderName: string;
  companyId: string;
  whiteLogo: string;
  darkLogo: string;
  loginBackground: string;
  useCompanyName: string;
  shortcutLink: string;
  nbf: string;
  exp: string;
  iat: string;
  iss: string;
  aud: string;
}