// src/share/enum/notification-type.enum.ts
export enum NotificationTypeEnum {
  NEW_USER = 'NewUser',
  MULTI_USER = 'MultiUser',
  FORGOT_PASSWORD = 'ForgotPassword',
  RESET_PASSWORD = 'ResetPassword',
  FAIL_CONVERSION = 'FailConversion',
  SUCCESS_CONVERSION = 'SuccessConversion',
  FAIL_UPLOAD = 'FailUpload',
  TWO_STEP_TOKEN = 'TwoStepToken',
  BECOME_TO_MEMBER = 'BecomeToMember',
  ACCEPT_REQUEST_TO_MEMBER = 'AcceptRequestToMember',
  CLIPPING_PROCESSING = 'ClippingProcessing',
  CLIPPING_READY = 'ClippingReady',
  ACCOUNT_ACTIVE = 'AccountActive',
  ASSET_ASSIGNED = 'AssetAssigned',
  MIGRATION = 'Migration',
  SHARE_LINK_DISABLED = 'ShareLinkDisable',
  PROJECT_USERS_EDIT = 'ProjectUsersEdit',
  ANNOTATION_EDIT = 'AnnotationEdit',
  COMMENT_ANNOTATION = 'CommentAnnotation',
  USER_EDIT_PROFILE = 'UserEditProfile',
  BECOME_TO_MEMBER_REQUEST = 'BecomeToMemberRequest',
  BECOME_TO_MEMBER_APPROVED = 'BecomeToMemberApproved',
}